pyproject.toml
[tool.poetry]
name = "cdk-lab-rds-glue-athena"
version = "0.1.0"
description = ""
authors = ["Violet Vivirand <purpleslovemail@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.6"
aws-cdk-lib = "2.24.1"
constructs = ">=10.0.0,<11.0.0"
cdk-nag = "^2.14.33"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "6.2.5"
ipython = "^8.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
